fix:test_test.py -> start.py
